split admin tables into components
